Split room state hooks out into separate files
